Fix duplicate scroll wheel handler
Jan-Willem van Aalst (OpenTopo) pointed out that one "roll" of the mouse
wheel changes the zoom level by two steps instead of one. This turned out
to be a bug in Heron MC, which I have reported here:
https://github.com/heron-mc/heron-mc/issues/459
Until this is permanently fixed upstream, this patch remedies the problem
for our specific situation (it's not a general fix for Heron MC).